print given number to console

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -7,7 +7,7 @@
 void print_number(int n)
 {
 	int divisor = 1;
-	long long abs_n = n;
+	long abs_n = n;
 
 	if (abs_n < 0)
 	{
